+++ Sources/VaporRestKit/Resources+CRUDs/ResourceRelation+Delete.swift
@@ -82,6 +82,8 @@ public extension RelationsController {
     }
 }
 
+//MARK: - Concurrency
+
 public extension RelationsController {
     func deleteRelation<Model, RelatedModel>(
         resolver: ChildResolver<Model, RelatedModel> = .byIdKeys,
